Join/Rejoin channels on successful connection
Previously `client.Join` was an explicit send of the join command to the
server. This meant that if your connection was dropped and you
reconnected you would not join any channels on successful reconnection.

This change updates that so a map of channels is maintained so that if
the connection is lost, upon successful reconnection all of the channels
are rejoined.

Due to the way join worked previously, it could be called before or
after `client.Connect`, so careful consideration was taken to ensure
that the same behavior is retained with this change.

Addresses issue #34