Added emote position extraction to emote parse.
This CL adds a new field called Positions to the Emote struct. Positions is a list of EmotePosition, which marks the start and end character of an emote, as sent by the Twitch API.

This allows for downstream consumers of this library to quickly replace emotes in text, without the need to parse/regex the message string.